just rename / move NotificationController and CoroutineIntentService
