+++ app/src/main/res/mipmap-ldpi/ic_notification.png
