+++ app/src/main/res/mipmap-mdpi/ic_notification.png
