+++ app/src/main/res/values/strings.xml
@@ -91,7 +91,7 @@
     <string name="confirmation_cancel_prev_download_title">"Stop the current search and scan here instead?"</string>
     <string name="confirmation_cancel_prev_download_confirmed">"Scan here instead"</string>
     <string name="confirmation_cancel_prev_download_cancel">"Continue current scan"</string>
-    <string name="notification_downloading">"Scanning for quests…"</string>
+    <string name="notification_syncing">"Syncing data…"</string>
     <string name="turn_on_location_request">"Location is off. Open Settings to turn it on now?"</string>
     <string name="no_location_permission_warning_title">"Location Permission"</string>
     <string name="no_location_permission_warning">"Location permission is needed to show your position on the map and scan for quests in your vicinity."</string>
@@ -705,7 +705,7 @@ Otherwise, you can download another keyboard in the app store. Popular keyboards
     <string name="quest_address_answer_no_housenumber_message1">The building was tagged as:</string>
     <string name="quest_address_answer_no_housenumber_message2b">Is this correct?</string>
     <string name="quest_maxspeed_sign_question">What defines the speed limit?</string>
-    <string name="notification_channel_download">"Download"</string>
+    <string name="notification_channel_sync">"Sync Data"</string>
     <string name="quest_playground_access_title">Is this playground publicly accessible?</string>
     <string name="quest_segregated_title">How are the footway and cycleway laid out here?</string>
     <string name="dialog_tutorial_upload">To upload your changes manually, use the button in the toolbar that looks like this.</string>
